requirements.txt
google-genai>=0.2.0
ddtrace>=2.0.0
datadog-api-client>=2.0.0
python-dotenv>=1.0.0

